🍓 test : profile 테스트 코드 수정 및 추가
- 프로필 테스트 코드 통합된 프로필 버전으로 추가
- 리뷰 엔티티 및 dto 수정

Related to #24

+++ src/main/java/com/mindmate/mindmate_server/review/domain/Review.java
@@ -39,7 +39,7 @@ public class Review extends BaseTimeEntity {
 //    @JoinColumn(name = "matching_id")
 //    private Matching matching;
 
-    private Integer rating;
+    private double rating;
 
     @Column(columnDefinition = "TEXT")
     private String content;
@@ -50,7 +50,7 @@ public class Review extends BaseTimeEntity {
     private String reply;
 
     @Builder
-    public Review(User reviewer, User reviewee, Integer rating,
+    public Review(User reviewer, User reviewee, double rating,
                   String content ) {
         this.reviewer = reviewer;
         this.reviewee = reviewee;
